Switch PKGBUILD to fetching git snapshot.
I haven't finished a full test build of this yet, but it fetches and begins building.  I'll run through a full build test tomorrow.  No config changes were made so the test is really to make sure all of the cp's and mv's still have paths that make them happy.