linux-odroid: build initrd image using mkinitcpio
Although it might look like an inessential feature, this initrd generation is
actually a fix to a real problem on Odroid U3. Early userspace is necessary for
specifying root file system by a persistent indentifier, such as LABEL or UUID
of a partition. On Odroid U3 board in particular, a persistent paritition
identifier is a necessity, not a consmetic issue.

On Odroid U3 board there are two *removable* block devices: SD card and MMC
module. If only one of the two is plugged in, then the one and only device is
assigned index 0. Otherwise, the MMC module is assigned index 1 and the SD card
index 0. Thus, the MMC module may be either at index 0 or 1 depending on
whether an SD card is plugged into the SD slot. As a result, any given U-Boot
configuration using device numbers can either boot the system when the SD card
is present OR when it is not present, but not in both cases.

This change is compatible with the current version of uboot-odroid. U-Boot
already looks for the ramdisk by default. The compiled-in behavior, the
'rdfile=' setting in uEnv.txt can only override the file name. This change
generates the image with the name that matches the default name in
uboot-odroid.