alarm/odroid-c1-libgl-headers: add conflicts for libump
The packages odroid-c1-libgl-headers and libump(-git) own the same
files in the file system (at least partially). This is easily
reproducable by having odroid-c1-lib-headers installed and trying to
install libump. The files in question are:
    /usr/include/ump/ump.h
    /usr/include/ump/ump_debug.h
    /usr/include/ump/ump_osu.h
    /usr/include/ump/ump_platform.h
    /usr/include/ump/ump_ref_drv.h
    /usr/include/ump/ump_uk_types.h

This fix is therefore needed to resolve conflicts before
actually trying to put files into the filesystem that
already exist (and thus quit with an error).