libcec 3.1.0 does not include Exynos-specific files by default.
As a consequence, Exynos devices (Odroid-XU for example) do not work properly (try running `cec-client` on an Odroid-XU, it won't work).
This acts as a workaround. With this patch, libcec compiles in Exynos-specific files so that any program relying on this library (Kodi for example) works. Try running `cec-client` with this patch, it works properly (as root).

See https://github.com/Pulse-Eight/libcec/issues/201