systemd: coredump: revert replacement of %p with %P
This reverts d1fcdcd87a4f0ffebfc98f404a7eae477dc3fd86
to fix systemd-coredump for kernels < 3.12.

The %P template param -- PID as seen by initial PID namespace, as
opposed to the PID namespace in which the process resides (e.g. a
container) -- is not supported in kernels < 3.12, which causes
systemd-coredump to get wrong argument list and interpret fields
incorrectly.

**Motivation**: Since manufacturers of platforms targeted by Arch Linux
ARM often stop updating the kernel (example: Odroid U3 frozen at 3.8),
it is useful to stay compatible with older kernel versions whenever this
is simple enough. Since most users are not using containers, but do want
working coredumps, the default should make the common case work, and
require one extra config step in the uncommon case when containers are
used. The current situation is the reverse.