Add getty.target to kodi systemd service
When running the kodi systemd service at boot and you play a movie with blackborders (or wait for the screenscaver dimm) you can see the console login still in the background. With a manual start of the service this does not happen. You need to wait for the getty to finish.