alarm: uboot-veyron: update version, add support for veyron speedy
0001-rk3288-set-CONFIG_SYS_TEXT_BASE-for-depthcharge.patch:
dropped because CONFIG_SYS_TEXT_BASE was moved to Kconfig in commit
278b90ce786f73faf29aa522d5d101e1da006378, replaced by:
0007-veyron-set-CONFIG_SYS_TEXT_BASE-for-depthcharge.patch

0001-rockchip-add-support-for-veyron-speedy-ASUS-Chromebo.patch:
My patch working its way upstream for the veyron speedy chromebook.

0002-rockchip-fix-incorrect-detection-of-ram-size.patch:
Needed for the 4GiB variant of the c201 as the amount of bytes of ram
are currently stored in a 0xffff_ffff sized variable and 4GiB in bytes
is 0x1_0000_0000, causing this variant to appear to have 0B ram.

0003-Revert-doc-fix-incorrect-usage-of-DT-node-unit-addre.patch:
Fixes boot on jerry.

0006-rockchip-speedy-Add-a-.its-file-for-chromium.patch:
Just an its file to make the coreboot bootable image file.

0008-dtoc-python2-compatibility.patch:
Updated to apply to u-boot-2018.07.

Signed-off-by: Marty E. Plummer <hanetzer@startmail.com>