core/linux-armv7: support flashing to chromebook internal storage
The current code assumes that the installation is done to partition
number 2, and consequently that the kernel is in partition number 1. I
installed Arch on an ARM Chromebook by first resizing the ChromeOS state
partition, and them using partitions 6 and 7 from the internal MMC for
kernel and rootfs, respectively.

This change makes the code generic: if the root partition is on
/dev/xxx(N), then it will flash (or suggest the user flashes) the kernel
to /dev/xxx(N-1).