Fix unapplied ARM linux kernel patches in PKGBUILD
PKGBUILD is wrong for aarch64 linux kernel:
The linux source code is downloaded as tgz and then patches are applied
with git apply.
It is not allowed to do so, and doing so lead to unapplied patches
without throwing error. So lots of the patches are not applied.

This patch use patch comand instead of git aply.

Signed-off-by: Abdel Benamrouche <abdel@unxp.fr>