Add patch to enable widevine on armv7h
Since chromium 78, extra flags were added to the chromium source code
to check for the target architecture and operating system in order
to decide whether or not to look for the widevine library.  This
meant that since that version it's not possible to use a chromeos
extracted widevine library anymore (which worked perfectly).

This patch adds "arm" and "linux" as a valid combination of arch and
os in the relevant gni file (i.e. build flags).  No modification
of actual C++ source code is required.
This brings back original functionality which had worked fine in the
past (chromium 78 and before).

This patch was tested with the official archlinuxarm distcc toolchain
and it does build on armv7h and aarch64 for chromium 86.
I've also checked against the git repository, and the patch will
apply cleanly to all the foreseeable versions in the future
(currently these are the latest development versions of 87 and 88 as
tagged in git).