alarm/kodi-rbp4 to 18.9-3
Calling the kodi-gbm target in ExecStart= was not actually the key to
fixing the unclean shutdown bug. Adding the ExecStop= calling killall
to the exact target was needed.

This commit restores the call to kodi-standalone to allow the
"RestartApp" built in to work again and adds the key ExecStop= with
killall.