alarm/kodi-rpi-git to 19.0b2.56808.0f75c91014-4
Properly set cortex-a72 optimizations for this package, see comment
in 0000-use-mcpu-avoiding-march-and-mtune.patch for more details.

Should be totally fine in doing so since this only RPi 4 can actually
make use of the HW decoding intended for v19 once it goes final and the
patchset is gets merged into ffmpeg (at some point).

In addition to that, fix a small bug in kodi.service, for details see:
https://github.com/graysky2/kodi-standalone-service/commit/909f274d6eaf011add6326b28b42fecb9123c7df