alarm/raspberrypi-bootloader to 20201215-1
This is actually a downgrade relative to 20201214-1 which
introduced a critical bug preventing booting all RPi 0-3
but not RPi 4 on which I tested the package.

Rather than adding an epoch= to the PKGBUILD, just bump the
pkgver but ship the last-good commit from upstream previously
provided by our 20201208-1 package.

Upstream bug report:
https://github.com/raspberrypi/firmware/issues/1511