added community/consul-1.9.1
The last committed version of consul was 1.5, when the upstream
community package was still relying on make targets. As of at least 1.9,
upstream packages now build via explicit calls to `go build` which is
less problematic in terms of how the upstream GNUmakefile makes
assumptions about amd4.

Changes from upstream:
- GOOS and XC_OSARCH can be safely omitted, the build process detects these
- A nodejs makedep is not present on armv6. Using another lts node
  release has been tested and confirmed to work after building the
  package with this change (on armv6, armv7, and aarch64)

I have built this on armv6, armv7, and armv8. I've defined a buildarch
to exclude armv5 since I can't definitively confirm it works on that
arch, but the rest build and run great.