extra/mesa: enable LTO
While LTO does increase build times and memory requirements for
the builder, it also results in a 20% score increase in
glmark2-es2-wayland on a RK3566 Quartz64 using the panfrost
driver.

I haven't seen the used memory during the build go above 2.5G so
I think we're good on that front as well.