core/linux-aarch64: enable Quartz64 A support
This enables a few config options to support the Quartz64 Model A
board by PINE64.

Notably, CONFIG_MOTORCOMM_PHY needs to be y instead of m to stop
the generic PHY driver from being loaded before the motorcomm one,
as the motorcomm chip has a zeroed out vendor ID.

PCI-e host support is enabled though the PHY for it is in 5.18 so
this won't work yet, but will be needed a release from now.