core/linux-armv7: Fix build in git
The PKGBUILD uses `git apply` [0] as a glorified `GNU patch` [1], which
works fine if you are not in a git repository since the file-paths in
the patch file are interpreted relative to the current working directory.
But it breaks horribly and in some cases even silently if you are
*somewhere* in a git repo, because the paths in the patch are then
interpreted relative to the git repo root.
I am not going to argue about the use of `git apply` here, but will
assume there are reasons for it and not touch it. But I am not the
first [2] and probably not the last that runs into this (undocumented?)
restriction. And in my eyes it is quiet a reasonable use-case to clone
this very git repo, and be able to build packages directly from it.
By setting the `GIT_CEILING_DIRECTORIES` [3] env var, you can "confine"
git to a specific directory so it won't be able to detect any maybe
existing git repo further up in the file-tree, and thus make it always
work in the simple "GNU patch" mode. It is "only" set to `$srcdir`
because you need one directory in between for it to work. It is scoped
to the `package` function by unsetting it at the end.
Alternatively you could specify the `--directory` flag on every
`git apply` invocation, but the env var seemed more practical to me.

[0] https://git-scm.com/docs/git-apply
[1] https://man7.org/linux/man-pages/man1/patch.1.html
[2] https://archlinuxarm.org/forum/viewtopic.php?f=60&t=15933
[3] https://git-scm.com/docs/git#Documentation/git.txt-codeGITCEILINGDIRECTORIEScode