core/linux-armv7: Add firmware failsafe check
A simple check at the end of the `package` function to exemplary check
one of the firmware files that are added with a patch. If that patch is
not applied correctly, that file is missing and the kernel build will
later fail. All this really does is catch an obvious error early.
The exit code was arbitrarily chosen just so it is unique.