added alarm/ffmpeg-rpi
New package is required for HW decoding with vlc-rpi.  Use -mcpu in C/XXFLAGS
which will limit use of this package to RPi3 and above which isn't a big
loss since HW decoding on RPi2 is pretty limited.  Commit also makes this
a split package with the new ffmpeg-rpi-bin providing the man pages and
executables stripped out in the main package to allow a full package of
ffmpeg with HW decoding should someone what that.