linux-* fix patching silently failing when in-tree
Patches were being applied with `git apply` despite the kernel tarball
not being a valid git repo. When building in-tree, git would get
confused with the repository root directory, and patching wouldn't be
done. The resulting package would therefore be incorrect.

This replaces all calls to `git apply` with calls to `patch`.