extra/chromium: optimize eu-strip building logic
Use elfutils.git maintained and passed by makepkg instead of cloning
at build time

Cloning inside the scope of build() is out of makepkg's control, and
adds a break point that's not even build-related. It also wastes
bandwidth as for each build the source of elfutils.git need to be
cloned freshly.

Let's update the PKGBUILD so makepkg would maintain the source of
elfutils.git, and the build.sh would use the .git passed by makepkg.
Even better, if builder has SRCDEST that's shared between a lot of
packages this would save more bandwidth.