linux-armv7: config fixes for veyron chromebooks
I assume the RK808 driver was accidentally disabled
when the configure option was split into RK8XX_I2C and SPI.
The RK808 driver is necessary to boot the veyron chromebooks.

Without the Rockchip PWM driver, there's a rather long
period of the LCD being all-white during boot (before the
driver is loaded).

Without the Rockchip DRM built-in, there's no way
for the average user (without a serial port - that is
not even enabled by default) to be aware of simple
issues with the rootfs, because there would be
no visual output before module loading.