Add new pkg: qt6-base-headless
This is a lighter weight option for some packages which do not need
the full blown qt6-base.  For example, qt6-base installs to to a total
of 121.64 MiB and pulls down 25 dependency packages whereas qt6-base-
headless installs to <25% of that size and pulls in only 2 other
dependency packages.