add package: extra/ruff
Similar to some other packages, uv needs to be built with
`JEMALLOC_SYS_WITH_LG_PAGE=16` to work on some aarch64 systems. Upstream
is doing this for their official builds
(https://github.com/astral-sh/ruff/pull/4247) and I confirmed this works
on my Apple M1.

This is a followup from
https://github.com/archlinuxarm/PKGBUILDs/pull/2089