core/linux-aarch64: Qualcomm X Elite updates for v6.16
The Qualcomm X Elite CPU frequency scaling was enabled in v6.16, this
depends on the the Qualcomm CPUCP mailbox driver, so enable this.

The Synopsys eUSB2 driver config changed in v6.16, select the new option
to ensure USB still works on a variety of Qualcomm platforms.

Provide olddefconfig value for CONFIG_RUST to keep config in sync.

Signed-off-by: Bjorn Andersson <bjorn.andersson@oss.qualcomm.com>