linux-aarch64: Enable BTF information
This enables BTF information for the linux-aarch64 kernel, making it
possible to use BPF applications that require this to run. The BTF
information is generated using pahole, so add that as a build-depend.

The change makes the kernel packages somewhat bigger, as reported by
Pacman when upgrading:

Package (2)            Old Version  New Version  Net Change

linux-aarch64          6.18-3       6.18-4        42,54 MiB
linux-aarch64-headers  6.18-3       6.18-4       457,94 MiB

Tested on a Traverse Ten64 machine.

Signed-off-by: Toke Høiland-Jørgensen <toke@toke.dk>