alarm/uboot-rock64 to 2025.10-1 and add variants
Rock64 V2 boards are known to have sub-optimal DRAM trace routing
therefore the memory frequency must be reduced for stability.

Other distros (armbian) have settled on 786MHz across all board
revisions but this is not 100% stable on my V2, hence an additional
variant of the package that reduces this further to 667MHz.

Since it's possible to install the bootloader to SPI flash on these
boards, also add variants for this configuration.